feat: use datasource-mongo instead of mongoose
