+++ src/commands/projects/create/sql.ts
@@ -50,16 +50,8 @@ export default class SqlCommand extends AbstractProjectCreateCommand {
     this.dumper = agentNodejsDumper;
   }
 
-  protected async generateProject(config: Config): Promise<void> {
-    this.spinner.start({ text: 'Creating your project files' });
-    const dumperConfig = {
-      dbConfig: config.dbConfig,
-      appConfig: config.appConfig,
-      forestAuthSecret: config.forestAuthSecret,
-      forestEnvSecret: config.forestEnvSecret,
-      language: config.language,
-    };
-    const dumpPromise = this.dumper.dump(dumperConfig);
+  protected override async dump(config: Config) {
+    const dumpPromise = this.dumper.dump(config);
     await this.spinner.attachToPromise(dumpPromise);
   }
