+++ src/services/dumpers/templates/agent-nodejs/common/env.hbs
@@ -10,7 +10,7 @@ NODE_TLS_REJECT_UNAUTHORIZED=0
 APPLICATION_PORT={{ appPort }}
 
 DATABASE_URL={{ dbUrl }}
-{{#unless isMongoose}}
+{{#unless isMongo}}
 DATABASE_SSL_MODE={{ dbSslMode }}
 {{/unless}}
 {{#if dbSchema }}
