--- src/services/dumpers/templates/agent-nodejs/typescript/models/index.hbs
@@ -1,17 +1,0 @@
-{{#each models as |model|}}
-import type { {{capitalise model.modelName}}Interface } from './{{model.modelFileName}}';
-{{/each}}
-
-import Mongoose from 'mongoose';
-
-{{#each models as |model|}}
-import { {{model.modelName}}Schema } from './{{model.modelFileName}}';
-{{/each}}
-
-const connection = Mongoose.createConnection(process.env.DATABASE_URL);
-
-{{#each models as |model|}}
-export const {{model.modelName}} = connection.model<{{capitalise model.modelName}}Interface>('{{model.modelName}}', {{model.modelName}}Schema, '{{model.collectionName}}');
-{{/each}}
-
-export default connection;
