--- src/services/dumpers/templates/agent-nodejs/typescript/models/model.hbs
@@ -1,17 +1,0 @@
-import Mongoose from 'mongoose';
-
-interface {{capitalise modelName}}Interface {
-{{#each fields as |field|}}
-  {{wsc field.name}}: {{#if (isObject field.type)}}{{>renderNestedInterface type=field.type level=1}}{{else}}{{toInterfaceType field.type}}{{/if}};
-{{/each}}
-}
-
-const {{modelName}}Schema = new Mongoose.Schema({
-{{#each fields as |field|}}
-  {{wsc field.name}}: {{#if field.ref}}{ type: {{field.type}}, ref: '{{field.ref}}' }{{else if (isObject field.type)}}{{>renderNested type=field.type level=1}}{{else}}{{field.type}}{{/if}},
-{{/each}}
-}, {
-  timestamps: {{timestamps}},
-});
-
-export { {{capitalise modelName}}Interface, {{modelName}}Schema };
