+++ test/services/dumpers/agent-nodejs/agent-nodejs.unit.test.ts
@@ -180,7 +180,7 @@ describe('services > dumpers > AgentNodeJs', () => {
               forestAuthSecret: 'aForestAuthSecret',
               hasDockerDbUrl: true,
               dockerDbUrl: 'host.docker.internal',
-              isMongoose: false,
+              isMongo: false,
             },
           );
         });
@@ -476,22 +476,7 @@ describe('services > dumpers > AgentNodeJs', () => {
 
             expect(context.fs.writeFileSync).toHaveBeenCalledWith(
               `/test/a${language.name}Application/package.json`,
-              expect.stringContaining('"@forestadmin/datasource-mongoose": "^1.0.0"'),
-            );
-          });
-
-          it('should add mongoose dependency', async () => {
-            expect.assertions(1);
-
-            const { dumper, context, defaultConfig } = createDumper(language);
-
-            defaultConfig.dbConfig.dbDialect = 'mongodb';
-
-            await dumper.dump(defaultConfig);
-
-            expect(context.fs.writeFileSync).toHaveBeenCalledWith(
-              `/test/a${language.name}Application/package.json`,
-              expect.stringContaining('"mongoose": "^7.8.6"'),
+              expect.stringContaining('"@forestadmin/datasource-mongo": "^1.0.0"'),
             );
           });
         });
@@ -607,229 +592,6 @@ describe('services > dumpers > AgentNodeJs', () => {
           await expect(dumper.dump(defaultConfig)).rejects.toThrow(
             'Invalid value for FOREST_SERVER_URL: "invalidUrl"',
           );
-        });
-      });
-    });
-
-    describe('when handling mongoSchema', () => {
-      describe('when schema is empty', () => {
-        it('should not write models nor index', async () => {
-          expect.assertions(1);
-
-          const { defaultConfig, dumper, context } = createDumper(language);
-
-          defaultConfig.dbConfig.dbDialect = 'mongodb';
-
-          await dumper.dump(defaultConfig);
-
-          expect(context.mkdirp).not.toHaveBeenCalledWith(
-            `/test/a${language.name}Application/models`,
-          );
-        });
-      });
-
-      describe('when schema is not empty', () => {
-        describe('when dbDialect is mongodb', () => {
-          it('should create models directory', async () => {
-            expect.assertions(1);
-
-            const { defaultConfig, dumper, context } = createDumper(language);
-
-            defaultConfig.dbConfig.dbDialect = 'mongodb';
-
-            await dumper.dump(defaultConfig, {});
-
-            expect(context.mkdirp).toHaveBeenCalledWith(
-              `/test/a${language.name}Application/models`,
-            );
-          });
-
-          describe('when schema does not have any models', () => {
-            it('should write only the index file', async () => {
-              expect.assertions(1);
-
-              const { defaultConfig, dumper, context } = createDumper(language);
-
-              defaultConfig.dbConfig.dbDialect = 'mongodb';
-
-              await dumper.dump(defaultConfig, {});
-
-              expect(context.fs.writeFileSync).toHaveBeenLastCalledWith(
-                `/test/a${language.name}Application/models/index.${language.fileExtension}`,
-                {
-                  models: [],
-                },
-              );
-            });
-          });
-
-          describe('when schema does have models', () => {
-            it('should write as many files as models', async () => {
-              expect.assertions(3);
-
-              const { defaultConfig, dumper, context, schemaSample } = createDumper(language);
-
-              defaultConfig.dbConfig.dbDialect = 'mongodb';
-
-              await dumper.dump(defaultConfig, schemaSample);
-
-              expect(context.fs.writeFileSync).toHaveBeenCalledWith(
-                `/test/a${language.name}Application/models/collectionA.${language.fileExtension}`,
-                expect.objectContaining({
-                  collectionName: 'collectionA',
-                }),
-              );
-
-              expect(context.fs.writeFileSync).toHaveBeenCalledWith(
-                `/test/a${language.name}Application/models/collectionB.${language.fileExtension}`,
-                expect.objectContaining({
-                  collectionName: 'collectionB',
-                }),
-              );
-
-              expect(context.fs.writeFileSync).toHaveBeenCalledWith(
-                `/test/a${language.name}Application/models/collection-c.${language.fileExtension}`,
-                expect.objectContaining({
-                  collectionName: 'collection-c',
-                }),
-              );
-            });
-
-            it('should compute a safe camel cased model name', async () => {
-              expect.assertions(2);
-
-              const { defaultConfig, dumper, context, schemaSample } = createDumper(language);
-
-              defaultConfig.dbConfig.dbDialect = 'mongodb';
-
-              context.strings.transformToCamelCaseSafeString.mockImplementation(
-                name => `${name}camelCased`,
-              );
-
-              await dumper.dump(defaultConfig, schemaSample);
-
-              expect(context.strings.transformToCamelCaseSafeString).toHaveBeenCalledWith(
-                'collection-c',
-              );
-              expect(context.fs.writeFileSync).toHaveBeenCalledWith(
-                `/test/a${language.name}Application/models/collection-c.${language.fileExtension}`,
-                expect.objectContaining({
-                  modelName: 'collection-ccamelCased',
-                }),
-              );
-            });
-
-            it('should pass adequate configuration', async () => {
-              expect.assertions(3);
-
-              const { defaultConfig, dumper, context, schemaSample } = createDumper(language);
-
-              defaultConfig.dbConfig.dbDialect = 'mongodb';
-
-              await dumper.dump(defaultConfig, schemaSample);
-
-              expect(context.fs.writeFileSync).toHaveBeenCalledWith(
-                `/test/a${language.name}Application/models/collection-c.${language.fileExtension}`,
-                {
-                  modelName: 'collection-c',
-                  collectionName: 'collection-c',
-                  fields: [],
-                  timestamps: true,
-                },
-              );
-              expect(context.fs.writeFileSync).toHaveBeenCalledWith(
-                `/test/a${language.name}Application/models/collectionA.${language.fileExtension}`,
-                {
-                  modelName: 'collectionA',
-                  collectionName: 'collectionA',
-                  fields: [
-                    {
-                      name: 'aField',
-                      ref: 'a-collection',
-                    },
-                  ],
-                  timestamps: true,
-                },
-              );
-              expect(context.fs.writeFileSync).toHaveBeenCalledWith(
-                `/test/a${language.name}Application/models/collectionB.${language.fileExtension}`,
-                {
-                  modelName: 'collectionB',
-                  collectionName: 'collectionB',
-                  fields: [],
-                  timestamps: false,
-                },
-              );
-            });
-
-            it('should should compute safe camel cased references', async () => {
-              expect.assertions(2);
-
-              const { defaultConfig, dumper, context, schemaSample } = createDumper(language);
-
-              defaultConfig.dbConfig.dbDialect = 'mongodb';
-
-              context.strings.transformToCamelCaseSafeString.mockImplementation(
-                name => `${name}camelCased`,
-              );
-
-              await dumper.dump(defaultConfig, schemaSample);
-
-              expect(context.strings.transformToCamelCaseSafeString).toHaveBeenCalledWith(
-                'a-collection',
-              );
-              expect(context.fs.writeFileSync).toHaveBeenCalledWith(
-                `/test/a${language.name}Application/models/collectionA.${language.fileExtension}`,
-                {
-                  modelName: 'collectionAcamelCased',
-                  collectionName: 'collectionA',
-                  fields: [
-                    {
-                      name: 'aField',
-                      ref: 'a-collectioncamelCased',
-                    },
-                  ],
-                  timestamps: true,
-                },
-              );
-            });
-
-            it('should compute kebab cased file name', async () => {
-              expect.assertions(2);
-
-              const { defaultConfig, dumper, context, schemaSample } = createDumper(language);
-
-              defaultConfig.dbConfig.dbDialect = 'mongodb';
-
-              context.lodash.kebabCase.mockImplementation(name => `${name}kebab_cased`);
-
-              await dumper.dump(defaultConfig, schemaSample);
-
-              expect(context.lodash.kebabCase).toHaveBeenCalledWith('collectionA');
-              expect(context.fs.writeFileSync).toHaveBeenCalledWith(
-                `/test/a${language.name}Application/models/collectionAkebab_cased.${language.fileExtension}`,
-                expect.objectContaining({
-                  collectionName: 'collectionA',
-                }),
-              );
-            });
-          });
-        });
-
-        describe('when dbDialect is not mongodb', () => {
-          it('should not write models nor index', async () => {
-            expect.assertions(1);
-
-            const { defaultConfig, dumper, context } = createDumper(language);
-
-            defaultConfig.dbConfig.dbDialect = 'postgres';
-
-            await dumper.dump(defaultConfig);
-
-            expect(context.mkdirp).not.toHaveBeenCalledWith(
-              `/test/a${language.name}Application/models`,
-            );
-          });
         });
       });
     });
